Working on the LSTM
Still troubleshooting why the LSTM model won't work. As far as I understand right now it has something to do with the shapes of the data. I'm training the model on 5112 days of data but only asking it to output 366 days of data. I need to reorganize the array that the results will be appended to.